requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
gunicorn==21.2.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
slowapi==0.1.9
httpx==0.26.0
google-generativeai>=0.3.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0 